+++ lib/drawyer.dart
@@ -4,59 +4,54 @@ class Account extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ListView(
-          children: <Widget>[
-            UserAccountsDrawerHeader(
-              accountName: Text("DSC TIU"),
-              accountEmail: Text("dsctechnoindiauniversity@gmail.com"),
-              currentAccountPicture: CircleAvatar(
-                child: Text("D",style:TextStyle(fontSize:30.0,)),
-                backgroundColor: Colors.red,
-              ),
-              otherAccountsPictures: <Widget>[
-                CircleAvatar(
-                  child: Text("A",style: TextStyle(fontSize: 15.0),),
-                  backgroundColor: Colors.pink,
-                ),
-                CircleAvatar(
-                  child: Text("B",style: TextStyle(fontSize: 15.0),),
-                  backgroundColor: Colors.purple,
-                ),
-                CircleAvatar(
-                  child: Text("C",style: TextStyle(fontSize: 15.0),),
-                  backgroundColor: Colors.green,
-                )
-              ],
+      children: <Widget>[
+        UserAccountsDrawerHeader(
+          accountName: Text("DSC TIU"),
+          accountEmail: Text("dsctechnoindiauniversity@gmail.com"),
+          currentAccountPicture: CircleAvatar(
+            child: Text("D",style:TextStyle(fontSize:30.0,)),
+            backgroundColor: Colors.red,
+          ),
+          otherAccountsPictures: <Widget>[
+            CircleAvatar(
+              child: Text("A",style: TextStyle(fontSize: 15.0),),
+              backgroundColor: Colors.pink,
             ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Primary"),
+            CircleAvatar(
+              child: Text("B",style: TextStyle(fontSize: 15.0),),
+              backgroundColor: Colors.amber,
             ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Social"),
-            ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Update"),
-            ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Promotion"),
-            ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Forum"),
-            ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Spam"),
-            ),
-            ListTile(
-              leading: Icon(Icons.account_circle,size: 40.0,color: Colors.red,),
-              title: Text("Trash"),
-            ),
-            Divider(height: 50.0,),
+            CircleAvatar(
+              child: Text("C",style: TextStyle(fontSize: 15.0),),
+              backgroundColor: Colors.green,
+            )
           ],
-        );
+        ),
+        Category(type: "Primary", iconImage: Icons.computer),
+        Category(type: "Social", iconImage: Icons.person),
+        Category(type: "Promotions", iconImage: Icons.label_outline),
+        Category(type: "Updates", iconImage: Icons.info_outline),
+        Category(type: "Forums", iconImage: Icons.message),
+        Category(type: "Bin", iconImage: Icons.delete)
+      ],
+    );
+  }
+}
+
+class Category extends StatelessWidget {
+  final String type;
+  final IconData iconImage;
+
+  Category({@required this.type, @required this.iconImage});
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      leading: Icon(
+        iconImage,
+        size: 30.0,
+      ),
+      title: Text("$type"),
+    );
   }
 }
