+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:layout_learning/drawyer.dart';
 
 void main() => runApp(
   MaterialApp(
@@ -21,33 +22,7 @@ class _LayoutAppState extends State<LayoutApp> {
         title: Text("Layout Learning"),
       ),
       drawer: Drawer(
-        child: ListView(
-          children: <Widget>[
-            UserAccountsDrawerHeader(
-              accountName: Text("DSC TIU"),
-              accountEmail: Text("dsctechnoindiauniversity@gmail.com"),
-              currentAccountPicture: CircleAvatar(
-                child: Text("D",style:TextStyle(fontSize:30.0,)),
-                backgroundColor: Colors.red,
-              ),
-              otherAccountsPictures: <Widget>[
-                CircleAvatar(
-                  child: Text("A",style: TextStyle(fontSize: 15.0),),
-                  backgroundColor: Colors.pink,
-                ),
-                CircleAvatar(
-                  child: Text("B",style: TextStyle(fontSize: 15.0),),
-                  backgroundColor: Colors.purple,
-                )
-              ],
-            ),
-            Category(type: "Primary", iconImage: Icons.computer),
-            Category(type: "Social", iconImage: Icons.person),
-            Category(type: "Promotions", iconImage: Icons.label_outline),
-            Category(type: "Updates", iconImage: Icons.info_outline),
-            Category(type: "Forums", iconImage: Icons.message)
-          ],
-        ),
+        child: Account()
       ),
       body: ListView(
         children: <Widget>[
@@ -72,24 +47,6 @@ class Identity extends StatelessWidget {
           title: Text("Client Email",style: TextStyle(fontSize: 20.0,fontWeight: FontWeight.bold)),
           subtitle: Text("Client Name",style: TextStyle(fontSize: 20.0,fontWeight: FontWeight.bold)),
         )
-    );
-  }
-}
-
-class Category extends StatelessWidget {
-  final String type;
-  final IconData iconImage;
-
-  Category({@required this.type, @required this.iconImage});
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      leading: Icon(
-        iconImage,
-        size: 30.0,
-      ),
-      title: Text("$type"),
     );
   }
 }
